add new feature (add money to balance using the credit card #55

+++ src/main/java/com/example/wayz/DTO/StudentDTO.java
@@ -32,6 +32,6 @@ public class StudentDTO {
 
     private Integer tripsLeft=0;
 
-    private Double balance=0.0;
+    private Integer balance=0;
 
 }
